Send queries to fallback during snapshotting
This commit ensures that while snapshotting is ongoing we send all
queries to fallback. leader_ready is a proxy for snapshotting being
done, as this gets set only when snapshotting is complete. The flow goes
as such:
1. A leader is elected.
2. The process of creating a leader causes us to
initiate snapshotting.
3. When we verify snapshotting is complete we
send a notification that updates the leader_ready bool to true.

There are a number of external requests that are relied on by the
replicator during snapshotting, so we have to be selective about what we
do and don't block during snapshotting. For instance, we would normally
want to block extend_recipe during snapshotting, except when it is
called by the replicator. The replicator will send a message to inidcate
that snapshotting is finished, which is updated in the leader
asynchronously. This means that the replicator may end up calling
extend_recipe before the leader has been marked as ready. The remedy
this, we can tell if an extend_recipe call came from the replicator
because it will have a replication offset attached to it. We can use
this to be more specific about how we block on snapshotting in
extend_recipe, for instance.

This commit tries to be thoughtful about when and where we block due to
snapshotting.

Change-Id: I96e14ddd783320887a47a4389c543c7496bb5c82
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1395
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Justin Miron <justin@readyset.io>