installer: Deploy RDS DB, if necessary
If the user has asked us to deploy to a new RDS database, do so after
deploying the Consul stack. This required adding a db_name field to the
rds_db in the deployment, and either loading it from the existing RDS
database or prompting for it as part of the set of prompts for
parameters of the database to create.

We then query the stack to load the ID of the database we just created -
since in the currently deployed version of the stack there's no output
containing that database ID we have to dig into the resources, but once
https://gerrit.readyset.name/c/readyset/+/1394 is deployed that can be
significantly simplified.

Change-Id: Ic5e03270ed98adc2f475e937c64be73d074a26c7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1396
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>