installer: Handle cloudformation stacks already existing
If the installer is trying to create a cloudformation stack with a name
that already exists, which since we don't save the deployment until
after the stack has finished being created could happen if the user
kills the installer while it's waiting for a stack to enter the
CREATE_COMPLETED status, previously we'd just fail with an AWS API
error. Now, we detect this, and use the already-existing stack as if
we'd just created it, including waiting for it to enter status
CREATE_COMPLETED if necessary.

Note that this *does* false-positive if the stack has status
DELETE_IN_PROGRESS, then returns an error once the delete eventually
finishes since the stack goes away - as a later improvement, it'd be
nice to wait for the existing stack to be deleted then recreate, but for
now this should be fine - the user can always re-run.

Change-Id: Id83212001934bdc03fe5e6155aef3dbb6b636e91
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1399
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>