installer: Deploy readyset mysql stack
Once we've done all the preparatory work, deploy the actual ReadySet
MySQL stack with all of the parameters from the previous steps, but
otherwise leaving the fields as the default values in the stack.

After this, I'll modify this to allow the user to *change* any of the
configuration in the stack - the plan being to actually download the
stack itself to figure out what those fields are, so that when we make
changes to the stack we don't have to change the installer quite as
much.

Change-Id: I3a3b75af7a2fae9c123e05b572a9f898ff553013
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1404
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>