ops/benchmark: supporting AWS IAM components for benchmarking
The purpose of this CL is to introduce the necessary IAM role and policy
for Buildkite agents to assume during executions of the automated
benchmarking logic.

For now, the role simply has permission to access resources in the
DevOps assets s3 bucket within the build account, however, in the
future, we'll likely elevate permissions to support CloudFormation-based
benchmark stacks as well. We can cross that bridge when we get there
though, of course.

Change-Id: Ie3d2cd26a4a660fb5c880cbb53f32cb3deb78bb4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1410
Reviewed-by: Harley Klein <harley@readyset.io>
Terraform-Plan-Review: Harley Klein <harley@readyset.io>
Tested-by: Buildkite CI