noria: Made `DataflowState` serializable.
In order to be able to store the `DataflowState` in the `Authority`, the first
step is to implement `Serialize` and `Deserialize` for the `DataflowState`
structure.
This commit adds the `#[derive(Serialize, Deserialize)]` annotation for the
relevant structures, to achieve that.

For the [`SqlToMirConverter`], since today we don't have a centralized MIR graph
(we should strive to try to use a petgraph::Graph or similar solution), the MIR
nodes all have references to their parent and child nodes.
Sadly, this means that automatic serde implementation (by means of using
`#[derive(Serialize, Deserialize)]` is not possible, because we would run into a
stack overflow when the serializing algorithm starts resolving the references,
getting into a never ending loop (parent -> child -> parent -> child...).
Because of that, we are forced to implement our own serialization and
deserialization.

Serialization:
First and foremost, we replace those [`MirNodeRef`]s with the identifiers of
each node, namely the [`MirNode::name`] and [`MirNode::from_version`] fields.
Secondly, we will skip serdes for the [`MirNode::children`] and
[`MirNode::ancestors`] fields, which will have to be added afterwards (since,
again, they are references).
A note here: we don't do the same for the [`MirNodeRef`] that might be present
in some [`MirNodeInner`], which derives the [`Serialize`] and [`Deserialize`]
implementations.
We let those fields be serialized as [`MirNode`]s, even if that's not very
efficient (we can improve this later). We are guaranteed that we won't be
entering an infinite loop, since we are skipping the serdes of the `children`
and `ancestors` nodes in the [`MirNode`] fields.
This relies heavily on the fact that we SHOULD NEVER EVER HAVE circular
references in the [`MirNodeInner`] node reference fields.

Then, we will have the following foreign fields in the serialized structure:
- all_nodes: This is a map (conceptually at least, since the keys are not
strings and we have to serialize it as a vector in the end), which has the tuple
(name, version) as keys (the ids of the nodes), and the [`MirNode`] (with empty
children and ancestors fields) as the values.
This contains ALL the nodes, meaning we have to traverse the whole graph,
starting from the ones defined in the [`SqlToMirConverter::nodes`] fields. This
will be our source of truth upon deserialization.
- children: Another (conceptual) map, again with the tuple (name, version) as
keys, and the list of children ids as values.

Also, the [`SqlToMirConverter::nodes`] field gets serialized as a vector of
tuples (name, version).

Deserialization:
We start by taking the `all_nodes` vector and transforming it back into a map.
We take this opportunity to transform the [`MirNode`]s into [`MirNodeRef`]s.
We also reconstruct the [`SqlToMirConverter::nodes`] maps simply by iterating
the `nodes` field, and using the (name, version) to look in the `all_nodes` map
for the references. We copy them and that's it, we have our `nodes` field.
Then we traverse the `children` field and we do something similar: we look in
the `all_nodes` for the actual parent node and each children node, and we call
`add_child` on the parent node and `add_ancestor` on the child node, cloning the
references.
Finally, we iterate the `nodes` map and we make sure to correctly set all the
references in any possible [`MirNodeInner`], by taking the [`MirNodeRef`] that
they store, getting the real [`MirNodeRef`] from the `all_nodes` using the
(name, version), and replacing the references.

Refs: ENG-838
Change-Id: Ibef7b974d20b9a1e43ef5f5aa323dc81c005a466
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1437
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>