clustertest: Assert the set of allowed intermediate results for queries.
In clustertests we use `query_until_expected` to wait until a query
returns the expected results. This is typically used in place of
sleeping for the propagation and executing a query which can be flaky.

This CL adds specifying what the allowed set of intermediate values
are, and short-circuiting if a result is returned that does not
match. This should help us catch correctness bugs.

Change-Id: I0d04829b687a9fb895e0d987af8933c8f543fe66
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1415
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>