installer: Allow modifying cloudformation stack config
Allow the user to modify the bits of the configuration for the readyset
cloudformation stack that *aren't* passed in by the rest of the stack
before actually deploying the stack.

To avoid too tighly coupling the installer to the exact configuration
parameters that are in the stack, this happens by *downloading* the
stack from the URL, parsing it, and using that to figure out which
config parameters are available and what their default values are.

Change-Id: I81495dc06bec53b9b80509af884bbff1d289421c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1414
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>