Granular locking during replication
This CL changes the way we do snapshotting by completely
avoiding taking a global lock (with or without flush) on
all of the tables at once, but instead acquires a lock
for each snapshotted table individually. After taking the
table lock, the binlog position is read for *each* table,
and a transaction is started for that table. After that
the lock for the table is released, and snapshotting takes
place in the transaction.

Since locks are acquired for very brief periods of time,
and individually for each table there is no downtime for
a production database, and we can now safely connect to a
live replica.

The astute reader will figure out that this way each table
snapshot is taken with a potentially different binlog offset,
but thanks to the catchup logic Griffin implemented, the
tables will magically synchronize from the binlog.

Change-Id: I309ebfe71ec780aeea21cd28111261c56d0f3a9f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1425
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI