treewide: Upgrade aws- crates to 0.3.0
Upgrade all aws rust sdk crates we depend on across the entire repo to
version 0.3.0.

No particular reason I'm doing this, except for a desire to generally be
on the latest version of things when possible - plus this consolidates
the versions of these crates we're depending on across the workspace
which should reduce build time.

Change-Id: Ibe8bd4b40d6a8ae224ccdb3134b1d21b2796fd80
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1427
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>