installer: Always prompt for database name
It seems like it's possible to have an RDS database without a db name
configured on the instance itself, so unwrapping here isn't valid to do.
Beyond that, however, I thought about this a little more and I actually
feel like it makes more sense to *always* prompt for the db name,
regardless of what's configured on the instance, to allow the user to
connect to a different db name than what was originally configured on
the db. If one *is* configured on the db, we use it as the default value
for the prompt so the user can just hit the enter key to accept it if
they don't want to change it.

Change-Id: Iec0ed49a7a289b2cd5f8cc1e418b0d635fae559b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1430
Tested-by: Buildkite CI
Reviewed-by: Alana Marzoev <alana@readyset.io>
Reviewed-by: Peter Farr <peter@readyset.io>