installer: Check and fix VPC attributes
Check that the VPC we're deploying with has both the EnableDnsHostnames
and EnableDnsSupport attributes enabled, and prompt the user to allow us
to automatically update the VPC if they're not.

Since the validate_vpc() function now validates more than just subnets,
this also changes it to be run unconditionally, rather than only if we
don't have subnets set - it's idempotent, so this should be fine to do.

Change-Id: If956f6566144f2d5af6a58bce11e1a2a0efe62a0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1432
Tested-by: Buildkite CI
Reviewed-by: Alana Marzoev <alana@readyset.io>
Reviewed-by: Michael Cronce <mc@readyset.io>