installer: Add Postgres support
Add support for deploying in front of an existing Postgres database,
which works very similarly to, and generalizes out a few functions from,
the mysql database support.

This also fills in the cloudformation template for deploying a *new*
postgres database, but the outputs of that stack are in a different
shape so it likely won't work - getting that all ironed out is going to
come next.

Change-Id: I2d425db251a5f3c0e61ca9b6e41547b7c9ea5d28
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1433
Tested-by: Buildkite CI
Reviewed-by: Alana Marzoev <alana@readyset.io>
Reviewed-by: Michael Cronce <mc@readyset.io>