build: Use builder stage instead of cargo chef
cargo-chef does not actually speed up the build in a way that we want as
it only simplifies downloading dependencies. What actually speeds up
our builds more is things like using cachepot to cache compilation
results.

This creates a new image which is a base image for builders to use. We
could eventually in the future integrate in things like cargo-chef again
at some point on top of it.

The psql image has also been broken for a bit so this fixes that in the
process of updating everything else.

Change-Id: I37c1c6edc8772a7956e1cf80518839d2b05edee6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1435
Tested-by: Buildkite CI
Reviewed-by: Ryan Cerniglia <ryan@readyset.io>
Reviewed-by: Sasha Fox <sasha@readyset.io>