benchmarks: Sleep so `setup` DDL propagates to Noria.
This adds a 5 second sleep at the end of the `setup` step
in benchmarks to give the DDL time to propagate to Noria.

Due to binlog propagation delay, the data created via setup in the
benchmarking infrastructure may not actually reach ReadySet by the
time we run `benchmark`. Which can either:
 1. Impact the performance of the benchmark if we are doing a ton
    of writes, or
 2. Mean the schema hasn't hit ReadySet yet and our queries go to
    fallback when they shouldn't.

Change-Id: I2576db19b95e651f124ba441a46f05337c8ba512
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1439
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>
Reviewed-by: Marcus Bastian <marcus@readyset.io>