noria: Storing and recovering from failure using [`DataflowState`].
A couple of changes happened here:
1. We are no longer using the [`Recipe`]s for recovery. Instead, we are
using the persisted [`DataflowState`] for that. This meant that the
[`Recipe`] could be greatly simplified by removing the `prior` field,
meaning that we only keep one [`Recipe`], which gets updated.
It is important to note that we lose the ability to know which migration
introduced which change, but that should be fine. This tradeoff is worth
 it as it allow us to deal with the graph instead of a list of
 migrations, where the later is mostly useful for appending new elements
 to the graph, but not for removing elements (i.e, dropping tables).
2. We are now being very explicit as to _when_ we update the
[`Authority`] with a new [`ControllerState`]: we do so whenever we call
[`DataflowStateWriter::commit`]. In fact, we persist the changes to the
[`Authority`] first, _and then_ we make the changes to the [`Leader`]
itself.
This guarantees that:
  a. All changes made to a [`DataflowState`] will automatically be
  persisted.
  b. If a failure happens during `commit`, we either managed to persist
  it before failing (meaning the change will be present at recovery) or
  it's completely lost.
3. We are adding a new method [`DataflowState::recover`], which will be
used upon recovery (duh) and will take care of configuring the
[`Materializations`] and sending all the necessary messages to
initialize the Domains in the available workers.
4. This might be a subtle change, but we are now cloning the [`Node`]s
before calling [`Node::take()`]. We need this because the `take()`
function essentially "empties" the [`Node`], which is something we
cannot do anymore, as the [`Graph`] and its nodes must stay as-is in the
 case we must recover.

There are a couple of leftovers that should be tackled in the future, to
 cleanup the code even further. These changes are not included here in
 order to minimize the size of this commit:
1. Remove the `version` from the [`Recipe`].
2. Remove the `take()` and `finalize()` methods from the [`Node`] (they
are always called together), and maybe implement a function in the
[`DataflowState`] to perform both actions together.

NOTE: The avg,group_concat,equal_filters.test was moved to
logictests/out-of-scope/ENG-923. See ENG-923 for more details.

Refs: ENG-838
Change-Id: I8819d929200284a6908af582a9eae9356682cb13
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1452
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI