persistent-state: Give more info about missing CFs
Rather than just unwrapping, give a message - in the form of an
`expect` in one case and an `unwrap_or_else` with a panic in another
case, when we can't find a column family in rocksdb. This should never
happen, but if it *does*, this gives us more to go on than just a None
unwrap.

Change-Id: Idf31b6b72d8758b912e31074eb9f847318b3e352
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1464
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>