Update deny_list to work for explicit migrations
The deny_list method on the query status cache was previously only used
by async migrations. In the case of explicit migrations, the behavior of
this generated list is different. Any query that hasn't been explicitly
migrated can be inferred to be "denied" (or in other words, proxied).

This updates the query status cache to be constructed with a migration
style (either async or explicit) which alters the behavior of the
deny_list method accordingly.

Change-Id: I83c93d2834e811b358e3b7b7b4f162469e425473
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1466
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>