noria/benchmarks: v1.0 of automated benchmarks in CI w/ docker-compose
Adding a new CI pipeline for running automated benchmarks of Noria in
Buildkite. This pipeline gets called in the main pipeline, and will
always run on `main` builds. The pipeline may also be invoked during
feature branch pushes by unblocking the build step within the main
ReadySet pipeline.

Included documentation on how the data directory for the MySQL container
(within compose) is generated so we may refresh the data or change it in
the future, whilst evading the MySQL 8.0 lower_case_table_name woes that
delayed this effort.

Added scripts to orchestrate running the benchmarks in CI, as well as to
run the benchmark binary inside of the new benchmark binary container.
There's also an accompanying Dockerfile to compile the benchmark binary.

For now, this CL will rely upon a static 600s delay within the Benchmark
compose server's entrypoint, in order to ensure Noria snapshots all the
appropriate tables as expected before the benchmarks begin. A follow-up
CL will add a new benchmark binary flag to ask that process to take
charge of awaiting snapshot completion.

Additionally, this process is simply running the IRL benchmarks since
there were some issues with running the other benchmarks. To get
something going sooner than later, we're options to run things as-is for
now.

The existing image build process automation needed to be reworked to
allow for forcing a release binary based image to be built, and to
override what container image tag is built and pushed to ECR.

It's worth noting that for v1.0 we will not have a Grafana or Prom
server for metrics to be visualized or emitted to. This will follow in
the future.

Change-Id: I7af989249f8bcd061ef0c4373386f9e9f41ed2ef
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1472
Reviewed-by: Justin Miron <justin@readyset.io>
Reviewed-by: Harley Klein <harley@readyset.io>
Tested-by: Buildkite CI
Terraform-Plan-Review: Harley Klein <harley@readyset.io>