ops/join: Fix handle_upquery for compound keys
I'm not entirely sure what the idea behind the previous implementation
of Join::handle_upquery was - but the *intent* is to split out the set
of provided keys into keys on the left, and keys on the right. This more
or less reimplements the function to do so correctly, and covers it with
a couple of unit tests for good measure.

Change-Id: I2f9cd1b0c3a03c45deb77399bf0d239256c42893
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1476
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>