errors: Remove UnmappableWorkerIdentifier.
At some point, we removed setting worker nodes to unhealthy as once
the authority declares you are unhealthy, a node can never become
healthy once again. When that happens, the worker is removed from
the controller's worker set. This will cause UnmappableWorkerIdentifier
errors to be thrown on messages, when instead we should be returning
WorkerFailed.

Since we can no longer tell the difference between
UnmappableWorkerIdentifier and WorkerFailed, I've opted just to
keep WorkerFailed.

Change-Id: Ie2a9dd23fff19f9d4d589ade0b207a251d88fdd8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1486
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>