buildkite-pipeline-rs: Don't test against main pipelines
The buildkite pipeline deserializer in buildkite-pipeline-rs is
incomplete, so there are valid pipelines that it won't successfully
parse - we might want to *write* some of these valid pipelines, and
expecting someone who's just working on build infrastructure to context
switch to debug a serde::Deserialize impl is a little unfair. Instead,
this just copies the *current* state of all of our pipelines into a
fixtures/ directory and tests against these - if we need to expand
support, we can do so intentionally.

Change-Id: Ic9f44754e52223f0c9020bdf7e6e691584b5b1e1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1489
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>