installer: Add wrapper script
Add a wrapper script for the installer (which is being publically
referred to as the "readyset orchestrator"), which is responsible for
detecting the host platform and then downloading the appropriate binary
from the website I'll be setting up to host both this script and the
binary.

The vast majority of the content of this script is lifted directly from
the equivalent script that Rustup uses. Since the script itself is going
to be distributed verbatim to end-users, the bits that are *verbatim*
taken from rustup are all annotated with a copyright header (per the
requirements of rustup's MIT license). Reviewers can probably skim over
those functions - think of them as inlined library code.

Change-Id: Idbfda1cf6c87e1b111e04a61d77d4ef76e0ab2c1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1491
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>