ops/substrate: Add cert+distribution for launch.readyset.io
Add a certificate, validated cloudfront distribution, and domain for
https://launch.readyset.io in front of the newly-created installer
bucket, to give us a nice domain to host the distributed assets for the
installer.

The DNS part of this is happening via a newly-added unofficial Vercel
terraform provider, which has been configured with the VERCEL_TOKEN
environment variable in the secrets bucket.

Change-Id: I51359d02887a0b4c243254eabc50c6e0cc0cae66
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1492
Terraform-Plan-Review: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>