logictests: Add failing test for compound-keyed topk
This returns no rows, but it should return 1.

I haven't done a *ton* of triage on what the actual issue is, but upon
first blush it looks like we're creating two separate replay paths to
the topk with two separate tags for the same set of columns but in a
different order. That could be a red herring, though

Change-Id: Id1c00bbb4c6eb90426f57751a4c1f9dd710a28c2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1493
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>