migration: Allow remove_node requests to fail for failed workers.
If a worker has failed, it has removed the nodes from its state.
As a result, allow remove_node requests to fail when sent to
failed workers. `send_to_healthy` will alos log a message that
we tried to send a message to a failed server.

This currently fails as part of our failure handling logic as we
(1) identify the set of nodes to remove from the graph as a result
of the failure, (2) tell all the domains that those nodes are on
to remove the nodes, including the domains on the worker that
failed!

This combined with adding aliases to table names fixes the
create_view_after_worker_failure clustertest. The leader
no longer panics during the migration after a worker failure.

Change-Id: Id02214b5fb72f847e01b76f5799129b4380cec67
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1495
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>