Add last_query field to Backend
This adds a last_query field to the Backend struct which stores metadata
about the last executed query. For now this only includes whether the
query was sent to fallback or not, but may be expanded in the future.

This field is reset to None for each incoming prepare, execute, or
query, and set accordingly in program flow.

This will later be accessible via a new `EXPLAIN LAST STATEMENT`
command, and will be tested using logictests with that command in a
future commit.

Change-Id: I1b624e4d88d8a486e7d18641ea3d10a0a2c1214d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1505
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>