Propagate placeholder literals to reader map
This change carries placeholder indices (set according to the order that
they appear in the SQL query) through the query graph to MIR to the
Reader node to the View. This CL does not add or remove any
functionality.

This new structure allows us to easy mapping of user supplied parameters
to their associated key columns in the reader node. Previously this was
done by matching the column and table names of placeholders to reader
key columns. However, this is not foolproof as there could be multiple
reader key columns for a single parameter, or reader key columns that do
not correspond to a column in a database table (e.g., in the case of
limit/order).

The mapping is stored as an ordered Vec<(placeholder_idx,
reader_key_index)>. The vector is sorted once in MIR and can be iterated
over in order in the adapter to determine the key_column_indices (and in
a future CL, which keys are ignored).

This CL does not cover mapping placeholders to reader keys in the case
of reuse. We will also need to implement a method to deconflict
placeholder to reader key mappings if we allow multiple functionally
equivalent queries to share reader nodes (e.g., queries that are the
same except for the ordering of parameters in their where clauses).

Change-Id: I8b7cf6b18714da70d4e4a5a96ebca7ca9bf8a473
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1553
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>