server: Add flag to allow range queries
Add a flag to the noria-server binary threaded down through to the MIR
config to allow creating range queries, for use in testing. This flag
basically mirrors the equivalent flag for allowing topk queries, the
comment for which has been updated as well since I was in there anyway.

Similarly to topk, once range queries are fully supported we should
remove this flag entirely.

Change-Id: I0010f3a6d49875560d4524bba38f4e51b9b21388
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1507
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>