clustertest: Test that writes survive worker failures.
This adds two clustertests, one that performs a write
during a worker failure and verifies that the write
is eventually queryable through Noria. The other is a long
running chaos tests that repeatedly fails workers while
updates are being performed.

Both are marked should_panic, as we currently drop writes
to workers that have failed.

Change-Id: Ifc725336d701c44ae83874cd684e2d2cf396f22e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1510
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>