noria: Simplified [`Migration`] and [`MigrationPlan`].
Today, we have a new, mutable copy of [`DataflowState`] being used to run
migrations.
This means that we don't actually need to be careful when migrating, as any
error will short-circuit the logic and avoid calling
[`DataflowStateWriter::commit`]. Hence, the changes get discarded.
This also makes the code cleaner, as we now hold and modify one structure (the
[`DataflowState`]), and we can tie a [`Migration`] and [`MigrationPlan`] to the
lifetime of the dataflow state.

For now, most of the logic is the same (or almost the same) as how it was, for
the sake of making this commit as small as possible.
In the future, we would like to effectively rethink the `plan` and `apply`
methods from the [`MigrationPlan`]:
- `plan` should take care of modifying the graph and adding all the new nodes
and domains (without actually sending or preparing any domain message).
- `apply` should send all the messages that we need to send in order to inform
the Workers/Domains about the necessary changes required for the migration to be
successful.

Refs: ENG-838
Change-Id: Id1b18bf69384b644c1cb21b5ee328623d672d2bc
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1508
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>