dataflow: Properly return Occupied in PartialMap::entry
PartialMap is supposed to return Entry::Occupied from the `entry` method
if given a key that is covered by a range that has been marked as
filled, even if no actual entry for that key exists in the underlying
btree map - this is documented behavior, and depended on by
materialization to actually insert records into filled partial
holes (since KeyedState::Insert pattern-matches on Entry). It wasn't,
though - instead it was just returning Entry::Vacant!

This commit updates `entry` to properly return `Entry::Occupied` when
given keys covered by the internal interval tree, implemented by
converting the `OccupiedEntry data type into an enum, which can either
contain a `btree_map::OccupiedEntry`, or (now) a
`btree_map::VacantEntry` which is filled with the `default` value
whenever a mutable reference to it is requested.

Change-Id: Iae439f89721dedfb0f6b4f0279fd61130c82f098
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1515
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>