dataflow: Support range queries through straddled joins
Update the handle_upquery method on Join to properly support range
remapping range queries for straddled joins into a pair of range
queries, one to each parent. This works rather similarly to the way
equal-queries work for joins, except that we need to jump through some
hoops to make sure the Bounds on either side of the range are the right
clusivity (Inclusive, Exclusive, or Unbounded).

This all comes with unit tests for the handle_upquery function in Join
itself, plus an integration test that makes a straddled range query
through a Join and checks that the right results are returned.

Change-Id: I9f8866ce01b5b80138d1b32f707238fcae846273
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1516
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>