benchmarks: Move `run_for` into benchmark.
As part of a set of changes to make every benchmark return results
that are understandable when manually run, and return easily
parsable results we can check in CI.

A step towards that is making benchmarks cleanly terminate. Previously
the set of long-running benchmarks supported terminating after a
duration set by `--run-for`. However, they uncleanly terminated
without a useful set of results.

This removes this unclean termination and moves the `--run-for` into
the benchmarks that actually used it: the multi-threaded benchmark.
The multi-threaded infrastructure now handles termination after
the run_for duration cleanly. A follow-up CL will allow the
multi-threaded benchmarks (and all benchmarks) return results.

Change-Id: I552707c88d655b7de19564a03aa21d81cccce403
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1518
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>