unbounded-interval-tree: Allow graphviz'ing individual nodes
Extract all of the logic for graphviz'ing the tree out into a method on
Node itself, and make the graphviz method for the tree just call that
function. This proved *extremely* useful when debugging balancing on
deletion, and I want to have it around in the future if we ever have to
modify this crate again.

Change-Id: I11709ccdc1c90ce810ff9d79a889d6accae38451
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1525
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>