+ scripts/aba-benchmark.sh and docker-compose overrides for local benchmarking
This provides a single entrypoint that handles running the benchmark
suite on two commits (twice on the first, once on the second - so-called
"A-B-A testing").  It does everything except clone the monorepo.

From start to finish, it ensures the monitoring stack is running; then
for A, B, and A again, it checks out the commit, builds the codebase,
starts Zookeeper/Consul/Redis/fallback using docker-compose, starts
noria-server and noria-mysql, runs the benchmarks, and finally tears
down Noria and the support stack.

Change-Id: Ia73b0168a51e9cd0be05927b37d992662bb08dc8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1530
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>