dataflow: Support range queries through unions
Support range queries made through unions, by buffering them in the same
way that we buffer point queries.

This actually turned out to be relatively straightforward - just change
the type in the key of Union::replay_pieces from Vec1<DataType> to
KeyComparsion, wrapping the whole thing in a struct so we can implement
Ord (we don't need to do anything *particular* with the ordering, since
the only *actual* ordered comparisons we do here are on either tags or
requesting_shard, so we just implement a basic lexicographic piecewise
comparison on point keys or ranges), then update all the call sites to
use the new types - including:

- Updating the bit where we check if a record is covered by a buffered
  key to use `KeyComparison::contains`, updating that method to be
  slightly more general and tidying up the docs along the way
- Taking records *out* of replays using a local btree and the
  RangeBounds implementation for KeyComparsion, rather than just looking
  them up by point key

Sicee this makes it work, this also de-ignores the failing integration
test for range queries through unions.

Change-Id: I67493ff4012049c3d77aca3bab6ad2e2f97119d7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1533
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>