query_generator: Allow generating range queries
Add a pair of new QueryOperations to query generator; one for a single
range parameter and one for multiple range parameters. These add a
single `column > ?` to the WHERE clause of the query, the same way that
the corresponding operations for point parameters do. To make sure we
generate values both in and out of the range, this also both adds the
ability to annotate when generating queries that a particular parameter
should always have a particular value, and then does that along with
setting the column generation spec for the column as part of adding the
parameter.

Also, this adds both "range_param" and "multiple_range_param" to the
list of small operations that can be generated with `noria-logictest
permute`, to allow for generating a full suite of logictests from
permutations of operations including range parameters now.

Finally, since we'll be running logictests that make range queries, this
also updates the logictest runner to set the allow_range_queries flag on
the server that it starts for verification.

Change-Id: Id1fa059e9b80d9c0e7671d12021fd131635c435d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1534
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>