persistent_state: Make range queries work correctly
There were a couple of issues with the way we were configuring rocksdb
that prevented it from working correctly with range queries:

* By default, RocksDB compares keys lexicographically bytewise - this
  can *coincidentally* work for *some* data types, but breaks down under
  more exhaustive testing - for example, when encoded via bincode
  negative integers are *greater* than positive integers.
* If prefix extraction, the HashLinkList memtable factory, and the
  PlainTable format are all enabled, RocksDB doesn't allow in-order
  iteration across different prefixes. The documentation just says
  "results are undefined", but I've seen anything from no results, to
  partial results, to correct results.

Fixing the former requires configuring a custom key comparator, which
since we don't currently support zero-copy deserialization of DataType
involves copying potentially very large strings, and the latter requires
disabling prefix extraction, swiching to the LinkList memtable factory,
and switching from the PlainTable format to the default SST table
format. All of these things have a reasonably significant performance
impact on point queries - I saw slowdowns as high as like 30% for the
microbenchmarks in persistent_state - so we don't want to do them
for *all* indices. Fortunately, we already have index selection threaded
through to all State impls, so we can just conditionally configure
things to optimize for point queries or range queries based on whether
we're being asked to create a HashMap or BTreeMap index. That's what
this commit does - essentially just extending index selection to apply
to the way we configure different column families in RocksDB.

When *reopening* the database, RocksDB errors out if you pass a
different set of column family params, so we need a way to determine the
parameters that the individual column families were created with in
order to setup the Options for the ColumnFamilyDescriptors. This is
happening here by just encoding it in the name and parsing it out from
the list of column family names in PersistentState::new, which is
a *bit* of a hack but seems to work fine for now. Note that this *is*
currently a backwards incompatible change, as reopening an existing db
with the old column family names will now error out.

This all also fixes 153 previously-failing logictests with range
queries.

Change-Id: I049869f82ce9ef2f8fe95a67769d0866f7a723f0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1536
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>